--- src/test/java/pt/tqs/g205/CommitsApplicationTests.java
@@ -1,15 +1,0 @@
-package pt.tqs.g205;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class CommitsApplicationTests {
-
-  @Test
-  public void contextLoads() {}
-
-}
